+++ Tests/UnitTests/DataLayer/TestQueryDb.cs
@@ -11,23 +11,35 @@ using Tests.EfClasses;
 using Tests.EfCode;
 using Tests.Helpers;
 using TestSupport.EfHelpers;
+using TestSupport.Helpers;
 using Xunit;
+using Xunit.Abstractions;
 using Xunit.Extensions.AssertExtensions;
+using Xunit.Sdk;
 
 namespace Tests.UnitTests.DataLayer
 {
     public class TestQueryDb
     {
+        private readonly ITestOutputHelper _output;
 
+        public TestQueryDb(ITestOutputHelper output)
+        {
+            _output = output;
+        }
 
         [Fact]
         public void TestDbQueryChildReadOnlyOk()
         {
             //SETUP
-            var options = SqliteInMemory.CreateOptions<TestDbContext>();
+            var options = SqliteInMemory.CreateOptionsWithLogging<TestDbContext>(log => _output.WriteLine(log.ToString()));
             using (var context = new TestDbContext(options))
             {
                 context.Database.EnsureCreated();
+#if NETCOREAPP3_0
+                context.ExecuteScriptFileInTransaction(TestData.GetFilePath("ReplaceTableWithView.sql"));
+#endif
+
                 context.Add(new Parent
                     {Children = new List<Child> {new Child {MyString = "Hello"}, new Child {MyString = "Goodbye"}}});
                 context.SaveChanges();
@@ -42,7 +54,7 @@ namespace Tests.UnitTests.DataLayer
                 children.Count.ShouldEqual(2);
             }
         }
-  
+
 
     }