+++ Tests/UnitTests/GenericServicesSetup/TestSqlErrorHandlerSync.cs
@@ -143,16 +143,20 @@ namespace Tests.UnitTests.GenericServicesSetup
         public void TestDeleteNoSqlErrorHandler()
         {
             //SETUP  
+            Book firstBook;
             var options = SqliteInMemory.CreateOptions<EfCoreContext>();
             using (var context = new EfCoreContext(options))
             {
                 context.Database.EnsureCreated();
                 context.SeedDatabaseFourBooks();
-                var firstBook = context.Books.First();
+                firstBook = context.Books.First();
                 var status = Order.CreateOrder("J", DateTime.Today,
                     new List<OrderBooksDto> { new OrderBooksDto(firstBook.BookId, firstBook, 1) });
                 context.Add(status.Result);
-
+                context.SaveChanges();
+            }
+            using (var context = new EfCoreContext(options))
+            {
                 var utData = context.SetupSingleDtoAndEntities<BookTitle>();
                 var service = new CrudServices(context, utData.ConfigAndMapper);
 
@@ -168,16 +172,20 @@ namespace Tests.UnitTests.GenericServicesSetup
         public void TestDeleteCatchSqlErrorTurnedOn()
         {
             //SETUP  
+            Book firstBook;
             var options = SqliteInMemory.CreateOptions<EfCoreContext>();
             using (var context = new EfCoreContext(options))
             {
                 context.Database.EnsureCreated();
                 context.SeedDatabaseFourBooks();
-                var firstBook = context.Books.First();
+                firstBook = context.Books.First();
                 var status = Order.CreateOrder("J", DateTime.Today,
                     new List<OrderBooksDto> {new OrderBooksDto(firstBook.BookId, firstBook, 1)});
                 context.Add(status.Result);
-
+                context.SaveChanges();
+            }
+            using (var context = new EfCoreContext(options))
+            {
                 var config = new GenericServicesConfig()
                 {
                     SaveChangesExceptionHandler = CatchUniqueError19