+++ Tests/UnitTests/Performance/CompareAgainstHandCoded.cs
@@ -14,6 +14,7 @@ using Tests.EfClasses;
 using Tests.EfCode;
 using Tests.Helpers;
 using TestSupport.EfHelpers;
+using TestSupport.Helpers;
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Extensions.AssertExtensions;
@@ -134,7 +135,7 @@ namespace Tests.UnitTests.Performance
 
             using (var context = new EfCoreContext(options))
             {
-                var utData = context.SetupEntitiesDirect();
+                var utData = context.SetupSingleDtoAndEntities<BookTitleAndCount>();
                 var service = new CrudServices<EfCoreContext>(context, utData.ConfigAndMapper);
 
                 using (new TimeThings(_output, "RunHandCoded ReadMany", 1))